Lists: A lot of definitions and properties
Most of these ended up unused because I shifted to NatSets at one point,
instead of using In to express subsets.

It didn't work out because in the end I had to prove that two proofs of
"being a subset" where the same, i.e. had to postulate extensionality
if I wanted to proceed.